Component Props Optimization
Change various components to use props covering less of the total
system. This alows Yew to optimize rendering better: Yew only needs to
re-render components whose props have changed, so using minimal props
rather than a reference to the entire gameboy reduces the re-rendering
needs of many components since their props don't always change when the
gameboy ticks, and additionally reduces the time needed to compare props
since it doesn't need to do a full `==` between two complete gameboys.

For memory, split the memory viewer into more components, including
separate components for the hex-editor-like view and for each line of
the hex-editor-like view. By using an OwningRef<Gb, [u8]> for the props
of these subcomponents, Yew is able to compare each memory segment to
see whether it needs to re-render that memory, then if that section has
changed, it is able to compare each line and avoid re-rendering lines
that haven't changed.

This is very important for memory, since there's a lot of memory (and so
a lot of table cells for Yew to re-render). When each hex-view table was
part of a single Memview component, any time any memory changed, Yew
would have to re-render every single cell in every memory table, and
perform a DOM Diff to decide what to actually update in the HTML. With
this component breakdown, it can re-render at both hex-section and
line-level granularity.

The performance improvement from modifying the memory view this way is
HUGE. From 96% of time being spent on re-rendering (split 60/40 between
rebuilding the tables and vdom diffing), to just 14% of time spent
re-rendering. This change makes having the memory view practical without
completely killing the performance (though performance still seems
somewhat poor? Unclear how fast it is working relative to real gameboy).