Use write_all not write for printing serial out.
`write` could return `Ok(0)` which the current code interprets as
successfully haven written the serial output character when that
actually means that nothing was written. IDK if this can actually happen
with stdout, but it's better to use `write_all` since that ensures
either everything gets written or an error occurs.