Build pausable CPU.
This new version of the cpu (gated behind feature `microcode`) uses a
stack-based virtual machine to implement instructions. This makes it
possible to pause instructions at various points in their execution,
most importantly at yield points.

Essentially this makes it possible to have a state-machine for each
instruction without customizing the state machine. Instead the state is
"where in the microcode for this instruction am I?". Variables used by
instructions are just stored on a stack, which turns out to work very
well for most instructions!

The old CPU `tick` function is renamed to `run_single_instruction` to
make it clearer that it always runs a whole instruction and uses
`ctx.yield1m()` to perform concurrent processing. When using feature
`microcode`, the `run_single_instruction` function remains available,
but is implemented in terms of microcode. Whenver the microcode
encounters a yield, `ctx.yield1m()` gets called automatically.

When using the `microcode` feature, a new `tick()` function is available
which advances the CPU until the next yield, and a couple other ticking
operations are possible.

If this CL is adopted fully, it will make the `ctx.yield1m()` function
unnecessary.