Break up IORegs and MemMappedIo.
Take the useage-specific registers like Serial Regs, Timer Regs, Apu
Regs, and Ppu Regs and move them to separate types defined alongside the
thing that they are the registers for, and replace the many-fields in
MemMappedIo with single fields of the corresponding mapped Regs type.

Delete the IoRegs type entirely and replace it with having each custom
context have accessors for the usage-specific regs, e.g. PpuContext
supplies ppu_regs() and ppu_regs_mut() for accessing those subsets of
the registers.

Add macros (both procedural derive macros and a macro_rules macro) for
deriving certain common types of MemDevice and change many of the fields
of these register types to one-byte MemDevices which are just mapped to
the correct address.