Allow read/write of blocks of [u8] in memory.
Add methods to MemDevice to read/write an entire block of memory in one
step. The slice of bytes read must fit entirely within the MemDevice
without wrapping. To facilitate reads/writes with wrapping, create a new
type called RootMemDevice which can read any u16 address and which
allows reading to slices which overflow/wrap.

Using this, add a feature to read/write arbitrary types to/from memory
and use that feature for a few places that want to read multiple bytes,
e.g. u16 and Object.