Make audio-based timing work on linux.
Also clean up some of the API for getting audio and frame updates from
the emulator so that they aren't returned from the tick function.

Add double-buffering to the emulator frames so that the last frame is
always cheaply accessible from the gameboy even when the next frame is
rendering. This makes it possible to wait to actually render in case the
emulator internally produces more than one frame in the middle (as might
be the case when you use a fast-forward mode, for example).